+++ Release notes.md
@@ -1,3 +1,13 @@
+## 1.1
+
+- iOS wrapper upgrade to [2.10.0](https://github.com/BlinkID/blinkid-ios/releases/tag/v2.10.0)
+
+### API changes
+
+- Removed: `AddGerMrzRecognizer`
+- Deprecated: `AddBarDecoderRecognizer`, `AddZXingRecognizer`
+- Added: `AddAusPassportRecognizer`, `AddAusIDCombinedRecognizer`, `AddBarcodeRecognizer`, `AddCroCombinedRecognizer`, `AddCzCombinedRecognizer`, `AddGerIDBackRecognizer`, `AddGerIDOldRecognizer`, `AddGerPassportRecognizer`, `AddGerCombinedRecognizer`, `AddSerbIDCombinedRecognizer`, `AddSlovakCombinedRecognizer`, `AddSlovenianCombinedRecognizer`, `AddSingaporeCombinedRecognizer`, `AddSwissPassportRecognizer`, `AddVinRecognizer`, `AddMrtdCombinedRecognizer`
+
 ## 1.0
 
 - First official version.